Allow '/' character in variable names
This improves error messages when the user mistypes built-in operations, like
`List/buold`.  Before this change that would lex as `List` followed by `/build`
and lead to a confusing error saying that the `/build` file was missing.  Now
it lexes as a single token and reports that `List/build` is an unbound variable.