Fix `InterpretOptions` to affect nested records.  Fixes #33
Consider the following program which marshals a Dhall record into Haskell:

```haskell
{-# LANGUAGE DeriveAnyClass    #-}
{-# LANGUAGE DeriveGeneric     #-}
{-# LANGUAGE OverloadedStrings #-}

import Dhall hiding (auto)

import qualified Data.Text.Lazy

interpretOptions :: InterpretOptions
interpretOptions = defaultInterpretOptions
    { fieldModifier = Data.Text.Lazy.dropWhile (== '_') }

data GitRepo = GitRepo
    { _host :: Text
    , _repo :: Text
    } deriving (Generic, Interpret, Show)

data BoxConfig = BoxConfig
    { _userName        :: Text
    , _dotfilesRepo    :: GitRepo
    } deriving (Generic, Interpret, Show)

main :: IO ()
main = do
    x <- Dhall.input (autoWith interpretOptions) "./config"
    print (x :: BoxConfig)
```

The above program is a common pattern when mixing Dhall records with lenses
(which typically prefix the original fields with "_").

Before this change, the above program expects a record of the following type:

```
{ userName : Text, dotfilesRepo : { _host : Text, _repo : Text } }
```

Note that the sub-record ignores the `InterpretOptions` and incorrectly includes
the underscore in the expected field names.

This change fixes the derived `Interpret` instance for records to correctly
thread `InterpretOptions` to nested records.

This required a change to the `auto` method of the `Interpret` class to
accept `InterpretOptions`, otherwise there is no way to supply the
`InterpretOptions` record to sub-records.  This modified `auto` method is now
called `autoWith`.  I still include an `auto` function that uses
`defaultInterpretOptions` consistent with the old behavior.