Fixed an issue with normalizeWith
Turns out that the previous `normalizeWith` didn't actually normalize.
For example, if `min` is a custom function, then the expression
`min (min +1 +2) +3` would have normalized as `min +1 +3`.

This is now fixed by

a) Allowing `CtxFun` to return `Maybe (Expr s t)` so that normalization
   can decide whether to keep normalizing or not.
b) Applying the Ctx fun only to `App`. Luckily, this is also the
   sensible thing to do in general. Having a context which freely edits
   the ast *anywhere* would be scary.