Revert "Overload `(++)` to work on `List`s.  Fixes #68 (#81)"
This reverts commit 8c340c12ddfd78e9dbfbd16013c00799a8f19895.

The reason for reverting this change is that using the same
operator name for appending `Text` and `List`s complicates
compilation to other languages.  For example, the `dhall-to-nix`
compiler currently translates `TextAppend` to `+`, but if you
consolidate `List` and `Text` append under a single `Append`
constructor then the translation to Nix is harder: you'd have to
do type inference as part of the translation process to know which
Nix operator to translate to.