Fix parsing single quotes in string literals
Fixes #96

The following Dhall expression would fail with a parsing error:

```haskell
"foo'bar"
```

... even though Dhall is supposed to allow single quotes in
double-quoted string literals.

The root of the problem was that Dhall was reusing the
`Text.Parser.Token.characterChar` parser to parse the body of the string
literal.  However, that's a parser designed for parsing the interior of
a Haskell *character* literal and not a Haskell *string* literal.  This
means that, for example, the parser does not accept just a single quote
for the same reason that Haskell character literals do not (i.e. `'''`
is not a valid Haskell character literal)

The solution is to just hand-write a parser instead that parses most of
the equivalent Haskell functionality

Note that this is a breaking change because it means that Dhall no
longer supports all the escape sequences that Haskell does.  However, in
practice this is not a huge loss since Dhall provides raw multi-line string
literals using "double-single-quote" string literals.