Left factor several parsers
This eliminates or narrows the scope of a large number of `try`s in the
parsing logic so that Dhall's parser does not need to backtrack as much.
The effect on performance for the example in #108 is minor at just 4%
(1.392 s → 1.340 s), however this change is still worth doing because
left factoring the grammar improves locality of parsing error messages

The changes included in this revision are:

* Reducing the scope of the `try` in `exprA4` to only backtrack until
  parsing the `arrow` symbol.  The parser can safely commit once the
  `arrow` is parsed
* Changing `exprB1` to only parse empty lists and optional values.
  Non-empty lists fall back to being parsed by `exprF` without the
  annotation.  This means that the parser can safely commit after
  parsing one list element
* Removing backtracking from `exprD`, which in turn requires enabling
  backtracking for the `exprIntegerLit` parser.  More generally,
  any primitive parser that shares a first character with an operator
  has to be wrapped in `try`. However, backtracking is cheap if it
  is limited to a single primitive parser
* Record types and literals are combined into a single non-backtracking
  parser
* Union types and literals are combined into a single non-backtracking
  parser

This also requires a small change to one of the Prelude examples and
tests.  This change causes a list to be now parsed without an
annotation, which in turn causes the result of normalization to not
include an annotation