Optimize `Natural/fold` and `List/fold` (#143)
This change modifies the behavior of `Natural/fold` and `List/fold` to
intelligently switch between a lazy or strict accumulator based on the
type of the accumulator

In some cases you want a strict accumulator, such as the following example
code:

```haskell
Natural/fold +3000 Text (λ(x : Text) → x ++ "!") ""
```

With a strict accumulator this runs almost instantly and without a strict
accumulator this takes a very long time to complete

In other cases you want a lazy accumulator, such as the following example:

```haskell
let generate = ./Prelude/List/generate
in  generate +6000 Natural (λ(x : Natural) → x)
```

This runs significantly more quickly with a lazy accumulator because the
accumulator is a list that grows on each iteration of the loop.  If the
accumulator is strict then you have to normalize the entire list on each
iteration which results in quadratic time complexity.  A lazy accumulator only
incurs linear time complexity

The solution is to modify each fold to take advantage of the fact that Dhall
knows the type of the accumulator without doing any type inference since the
type is passed as an argument to the fold.  Dhall inspects the type of the
accumulator to determine whether or not the fold should be strict:

* if the accumulator type has an upper bound on its size then be strict
* otherwise be lazy