Fix type-checking bug that permits infinite loop (#153)
* Fix type-checking bug that permits infinite loop

Fixes #151

Dhall is based on a pure type system that has the following rule:

    Γ ⊢ A : s
    ────────────────
    Γ, x : A ⊢ x : A

    Γ ⊢ b : B   Γ ⊢ A : s
    ────────────────
    Γ, x : A ⊢ b : B

... which both enforce that all terms in the context have to be type-checked
themselves before you can use the context to type-check other terms

You can introduce type-checking bugs if you don't type-check terms before adding
them to the context.  The reason why is that the type-checking logic often
normalizes types retrieved from the context and this normalization is not safe
if the terms have not been type-checked first.  Ill-typed terms can introduce
infinite loops when normalized, such as the following term:

    (λ(x : A) → x x) (λ(x : A) → x x)

Dhall's type-checking logic had exactly this issue before this change due to
not type-checking terms added to the context.  This change fixes the problem
and also adds a regression test to prevent this problem from recurring.

* Add missing files