Change Dhall's integrity check to use an "intensional" hash (#167)
This updates the recently added support for integrity checking to hash the
final normalized result instead of hashing the text of the import.  This gives
much stronger guarantees about the safety of the imported value, since the
hash now reflects changes to the expression's semantic value

This also lets you safely refactor imported values without changing the hash so
long as the refactor is behavior-preserving.  The hash provides a quick and
dirty protection against human error when used in this way