Use `Prelude/` prefix in tutorial examples (#184)
Fixes #59

The recent refactor of the Dhall grammar made `List/map` no longer a valid
identifier because the parser interprets it as `List /map`.  The grammar can't
be easily fixed to accept this variable name because:

* the normal way you'd fix this is to use a lexer with a longest-match rule
* however, Dhall cannot use a lexer due to nested block comments and string
  interpolation

So this sort of identifier name is no longer supported under the new grammar.
The work-around is to prefix the name with something that is not a built-in,
such as `Prelude/List/map`.

This change updates the tutorial to use the `Prelude/` prefix work-around so
that people don't accidentally run into this issue by following the tutorial.