Improve the speed of file-only imports
This overhauls `exprFromPath` (and every function downstream of that) to
speed up imports when no URLs are imported.

Previously, Dhall would always demand a `Manager` on the very first
import, whether or not Dhall actually needed to fetch a remote URL.
Fetching this `Manager` adds an unusually long overhead (~1 s) to
programs that only needed to fetch local files.

Now the `Manager` is only requested as late as possible (i.e. right
before the first URL import)