Fix infinite loop when pretty-printing `constructors` keyword (#211)
Fixes #209

The `prettyExprD` function would only transition to the `docs` loop if the
outermost constructor at precedence level "D" was an `App` constructor.  This
meant that the following code would pretty-print correctly:

```haskell
(constructors x) y
```

... but this code would not pretty-print correctly and would instead trigger
an infinite loop:

```haskell
constructors x
```

The reason for the infinite loop is that the pretty-printing logic assumes that
the pretty-printer will make progress for at least one precedence level in order
to terminate.  However, to the above issue the `Constructors` keyword would not
make progress at precedence level "D", which triggered the infinite loop.

The fix is to update `prettyExprD` to transition to the `docs` loops for both
the `App` and `Constructor` constructors.