Improve error messages for list elements with the wrong type
Fixes #230

The type error for the list now returns the expression for the mismatched
element instead of the entire list.  This improves the source span
displayed to the user to focus on the invalid element.

Since the source span now matches the invalid element, there is no need
to redisplay the same element in the detailed error message.