Fix pretty-printing of multi-line strings
Before this change the `dhall` executable would fail on the following input:

```bash
$ dhall <<< 'λ(x : Text) → "\n" ++ x'
∀(x : Text) → Text

»SFail« must not appear in a rendered »SimpleDocStream«. This is a bug in the layout algorithm! Please report this as a bug
CallStack (from HasCallStack):
  error, called at src/Data/Text/Prettyprint/Doc/Render/Util/Panic.hs:13:21 in prettyprinter-1.1.1-1CDqnG9d6HQ5GZzz2F5LpU:Data.Text.Prettyprint.Doc.Render.Util.Panic
λ(x : Text) → ''
```

This is because the `prettyprinter` library requires that any `hardline`
is guarded (using a `flatAlt`) by a `hardline`-free alternative.
However, the pretty-printing code for multi-line strings was forcing a
hardline without any fallback, triggering the above panic.

The solution is to fall back to a double-quoted string, which is free of
hard lines.