Fix output rendering (#280)
This change fixes two issues with output rendering

The first was the following panic in the prettyprinter library when
calling `Data.Text.Prettyprinter.Doc.Render.Terminal.renderIO` on an
expression without any annotations:

```
$ dhall <<< '1' > stdout.dhall
Integer

Peeked an empty style stack! Please report this as a bug.
CallStack (from HasCallStack):
  error, called at src/Data/Text/Prettyprint/Doc/Render/Util/Panic.hs:38:15 in prettyprinter-1.1.1-1CDqnG9d6HQ5GZzz2F5LpU:Data.Text.Prettyprint.Doc.Render.Util.Panic
```

The fix is to use `Data.Text.Prettyprinter.Doc.Render.Text.renderIO`
instead when the expression has no annotations.

The second issue is that the color detection was not correctly
working for `stderr`, meaning that this:

```
$ dhall <<< '1' 2> stderr.dhall
```

... would write escape codes to `stderr.dhall`.  The fix is to
separately check if `stderr` supports color or not when writing
to `stderr`.