Don't pretty-print header for `dhall` executable
The main reason for this change is stylistic (`dhall` doesn't need to
preserve the header because it's not filling the same purpose as
`dhall-format`)

The other reason for this is that `renderIO` was displaying the header
for both the type and the normalized expression.  Rather than complicate
the code to fix this I felt it was easier to just remove the
header-printing logic.