Sort fields when type-checking (#301)
This change ensures that the fields of a record or union are sorted
before comparing for equality.

I thought the old code was sorting the fields, but later learned that
neither `Data.HashMap` nor `Data.HashMap.Strict.InsOrd` guarantee any
ordering on the fields in their conversions and they have no `Ord`
constraints in their API.  Similarly, the `toSortedList` function had
no `Ord` constraint before this change.

However, what's suprising is that I could not devise a failing test case
despite the suspicious lack of sorting.  So it may be that there was
some sorting going on under the hood that I wasn't aware of but I'm
still sorting the results just to be safe.