Fix equivalence check to match standard semantics
Fixes #299

The standard mandates that two terms that are α-equivalent and
β-equivalent should be equal for type-checking purposes, but the
implementation was not matching the standard.  Specifically, a
type with an anonymous function would never compare as equal to
itself.

The newly added regression test shows an example of where the previous
implementation was misbehaving, where `λ(x : Type) → List x` would
not compare as equal to itself before.  After this change it correctly
matches itself since it is (obviously) the exact same type.

This also adds a new `alphaNormalize` function which was implemented
along the way, which converts an expression to α-normal form using
De Bruijn indices.