A command to interact with local ipfs installation
Since dhall has some affinity with ipfs I added a dhall-repl
command to directly push expressions to ipfs.

Here is an example of use:

    ⊢ :ipfs h = (\(x:Text) -> x++x)

    Added with hash: Qmee1h5H7o8Lp2FQtRJAPNGgLKRdrUSGn5jPkW6NBBxxh8
    ⊢ h
    "Qmee1h5H7o8Lp2FQtRJAPNGgLKRdrUSGn5jPkW6NBBxxh8"

Then on a completely different machine:

    ⊢ https://ipfs.io/ipfs/Qmee1h5H7o8Lp2FQtRJAPNGgLKRdrUSGn5jPkW6NBBxxh8 "hello "
    "hello hello "
    ⊢

I find this command delightful, but I won't be offended if this doesn't get
merged.

This incurs dependencies on typed-process, directory and deepseq. Also,
it needs -threaded to work properly.