Rename `Path{,Mode,Hashed,Type}` to `Import{,Mode,Hashed,Type}` (#376)
This updates the data types to match the terminology used in the
standard.  Specifically, "path" specifically refers to a directory or
file, whereas "import" is a broader term encapsulating URLs and
environment variables.

This is a breaking change but an easy one to adapt to.  Most downstream
consumers of the API just use the `Path` type, and this change includes
a `Path` type synonym for backwards compatibility.