Drop support for importing directories (#384)
Dhall used to support importing a directory, which would default to
importing a file named `@` underneath that directory (similar to how
`default.nix` works in Nix).  This change drops support for that since
there is no plan to standardize that feature.  I view this as sort of
a mis-feature that Dhall inherited from Morte because it it confusing
magic that deteriorates error messages when importing files.  Support
for `@` also complicates the standardization of the import system, so
I prefer to just drop support for it to keep things simple.