Update import logic to more closely match standard (#390)
This is one step in a series of commits to update the Haskell
implementation to match the import semantics standardized in:

https://github.com/dhall-lang/dhall-lang/pull/127

This is a BREAKING CHANGE mostly due to restricting the grammar for
valid file paths and URLs.  However, in practice most existing code
should still work after these changes.

Specifically, this change:

* Updates the grammar for file paths and URLs
* Simplifies and cleans up the semantics for canonicalizing paths (and
  chains of paths)

The main difference in the grammar is that file paths and URL paths are
unified to have the same syntax and they are both more restricted than
before.

For example, the following URLs are no longer valid:

* `http://example.com` (missing a path to a file)
* `http://example.com/()` (invalid path characters)

... and the following paths are no longer valid:

* `/?#` (invalid path characters)

Note that this actually deviates from the standard in one way: file
paths no longer permit `?` and `#` characters (whereas the standard
allows them).  However, in the course of testing these changes I
realized that if file paths permit `?` and `#` characters and you unify
URLs to also permit them then you get an ambiguous grammar where a `?`
or `#` in a URL could be interpreted as either part of a path component
or part of a query parameter or fragment.

The semantics for canonicalizing paths is a straightforward translation
of the standard to Haskell code.