Update parser to match latest grammar
... as standardized in https://github.com/dhall-lang/dhall-lang/pull/161

Note that this doesn't include the `whitespace-chunk` rule added by the above
change.  The reason why is that the `whitespaceChunk` ends up failing due to the
whitespace having already been consumed by a previous `whitespace` rule.  I
could make the whitespace rule backtrack at the cost of possibly performance
degradation and worse error messages, but I chose not to because the Haskell
implementation currently doesn't suffer from the ambiguous parses that the
`whitespace-chunk` rule was originally intended to address.