Data instance for Expr.
Fixes #416.

Also adds some Template Haskell to be able to load & resolve a Dhall expression
at compile-time.

Without the `Data` instance, the AST had to be pretty-printed back to a `String`
in the Template Haskell, and in 1.14 imports now auto-normalize, which often
results in much larger ASTs, and therefore (at least with the current version)
_much_ longer parses after re-serialization. Adding the `Data` instance avoids
that.

The current implementation of the `Data` instance removes
`Data.Text.Lazy.Builder` in favor of using `Data.Text.Lazy` directly, this is
simply because the latter already has a `Data` instance. It’s probably
reasonable to instead have a non-derived `Data` instance for `Chunks` that
explicitly uses `to`/`fromLazyText`, but that approach was less mechanical, so I
punted on it.