Remove a slow toList call in a hot-spot.
toList from insert-ordered-containers does some work to ensure that
the elements do follow insertion order. In this instance, insertion
order doesn't matter: we only need to peek at each element to make
sure it is well-typed. If there are multiple type errors, it does not
matter which gets reported first here.

This is less of a win than the previous commit, but it is still a
significant performance improvement.