Fix some performance issues (#420)
* Speed up type-checking large expressions bound with let

In #69, `let` was changed to substitute in the value of the bound
variable. However, if that value was large (e.g., the `constructor` of
a 300-constructor union), this has a very adverse impact on
performance where that value has to be re-typechecked at each use
rather than once when it is bound, leading to #412.

We can do better. #69 only changed anything when a type is being
bound. So we can check the type of the type being bound and change
strategy depending on whether it is `Type` or `Kind`: if it is a
`Type` we can get away with treating it like the pre-#69 behaviour; in
other cases, we fall back to substituting the value in.

* Remove a slow toList call in a hot-spot.

toList from insert-ordered-containers does some work to ensure that
the elements do follow insertion order. In this instance, insertion
order doesn't matter: we only need to peek at each element to make
sure it is well-typed. If there are multiple type errors, it does not
matter which gets reported first here.

This is less of a win than the previous commit, but it is still a
significant performance improvement.

* Omit re-typechecking variables at use sites.