Fix stripping of leading whitespace from multi-line strings
Fixes #468

The code to strip leading whitespace was incorrectly computing the leading
whitespace due to incorrectly escaping the string.  For example, for the
multi-line string:

```haskell
''
 foo
bar
''
```

The code was escaping the newline separating `foo` and `bar` and therefore
treating the multi-line string as one long line for dedenting purposes.  This
causes the code to incorrectly deduce that there is only one line with one
leading space instead of two lines with a minimum of zero leading spaces.

Fixing the code to not escape the string resolves the problem.