Add options and command-line support for protocol versions
The immediate goal of this pull request was to add command-line support
for a `--protocol-version` flag so that users could pin to a specific
binary protocol versions tring to avoid a forced upgrade when they
install a newer version of the Dhall interpreter.

However, in the process I ended up cleaning up the `Dhall.Import`
`load*` family of functions by consolidating everything into two
functions:

* `load` - Exact same type signature and semantics as before for the
  happy path with sensible defaults

* `loadWith` - The ultra-customizable version where all the options are
  stuffed into the existing `Status` record for ease of extension later

I also fixed a bug along the way where `decodeWithVersion_1_0` and
`encodeWithVersion_1_0` were serializing and deserializing the wrong
version string.