Fix test failure
The Nix build sets `HOME=/homeless-shelter`, which leads to a test failure
since the test tries to create a "${HOME}/.cache" directory without first
checking if "${HOME}" is accessible.  This changes the `assertDirectory`
function to recursively check parent directories to avoid this issue, which
in turn fixes the test (since it gracefully creates no directories if none are
accessible).