Make semantic integrity check insensitive to field order (#552)
... as standardized in https://github.com/dhall-lang/dhall-lang/pull/223

Fixes #548

This changes β-normalization to sort fields so that semantic integrity
checks don't change when reordering fields

* Fix test failure

The Nix build sets `HOME=/homeless-shelter`, which leads to a test failure
since the test tries to create a "${HOME}/.cache" directory without first
checking if "${HOME}" is accessible.  This changes the `assertDirectory`
function to recursively check parent directories to avoid this issue, which
in turn fixes the test (since it gracefully creates no directories if none are
accessible).