Fix build of `benchmark:dhall-command` on OS X (#611)
Due to OS X's case-insensitive filesystem the build fails due to it
accidentally resolving the `Dhall.Main` import to `src/Dhall/Main.hs`.
The fix is to just narrow the source directory to that one benchmark
so that it doesn't pick up that file by mistake.  I also made the same
change to other benchmarks as a precaution.