A couple more tests around kind polymorphism.
One fails because `Sort` isn’t parsed correctly. I added `Sort` to the “obvious”
places in the parser, but that’s not cutting it. Not sure if it’s something like
trying to parse `Some` and then failing to backtrack or what.

Also made a small change to the parser tests, since previously they were all
getting grouped under “whitespace”, and I don’t think that was correct.