Run tests in CI
CI was running tests, but they weren't blocking merge because they weren't a
dependency of the `dhall` attribute (which is what gates the GitHub status).

This fixes that by adding `coverage.dhall` as a dependency (which does include
tests).  The other builds were not running tests since that is what happens
`coverage` defaults to `false`.